Update code for smzdm
