+++ .env.sh
@@ -1,0 +1,12 @@
+#/bin/bash
+
+## source ./env.sh && python main.py
+
+export COOKIES=''
+
+## 钉钉
+export DINGTALK_ROBOT_TOKEN='' 
+export DINGTALK_ROBOT_SECRET=''
+
+## ServerChan
+export SERVERCHAN_SENDKEY=''
