+++ .gitignore
@@ -1,3 +1,3 @@
 __pycache__
-*/__pycache__
+**/__pycache__
 env.sh
