--- config.py
@@ -1,33 +1,0 @@
-"""
-请求头
-"""
-DEFAULT_HEADERS = {
-        'Accept': '*/*',
-        'Accept-Encoding': 'gzip, deflate, br',
-        'Accept-Language': 'zh-CN,zh;q=0.9',
-        'Connection': 'keep-alive',
-        'Host': 'zhiyou.smzdm.com',
-        'Referer': 'https://www.smzdm.com/',
-        'Sec-Fetch-Dest': 'script',
-        'Sec-Fetch-Mode': 'no-cors',
-        'Sec-Fetch-Site': 'same-site',
-        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.131 Safari/537.36',
-        }
-
-
-"""
-调试用 COOKIE
-"""
-TEST_COOKIE = ''
-
-
-"""
-调试用 SERVERCHAN_SECRETKEY
-"""
-SERVERCHAN_SECRETKEY = ''
-
-"""
-钉钉群机器人通知
-"""
-TEST_DINGTALK_ROBOT_TOKEN = ''
-TEST_DINGTALK_ROBOT_SECRET = ''
